Work around a floating IP bug in Neutron.
It's possible to create a floating IP in Neutron that has a v6 fixed IP and
a v4 floating IP (or vice-versa).  We had an employee do this while testing,
and it made pf on his router angry.  This commit adds a rule to ignore these
types of floating IP's (because they don't make any sense).