Sync requirements with stable/kilo GR for release
This syncs requirements.txt and test-requirements.txt with
global requirements in preparation for cutting our kilo
release.

This renames test_requirements.txt -> test-requirements.txt so
it can be managed by GR.

Also, this drops gunicorn as something we track in requirements.txt
This is really an external requirement and not a python requirement,
as nothing actually imports it.  This can be installed externally via
the appliance builder glue or devstack.

Change-Id: I74bf5de7d69f7ec6e4e0a3989178addd20aa0468