Adds support for NFS-based developer environment
This adds some optional stuff to the akanda element which will
run akanda-appliance code from an NFS server, allowing devstack to
export its code repo directly to running service VMs on boot. This
makes it easier to hack on the appliance without the need to rebuild
and republish the image and reboot on every change.

Depends-on: I3f5eca88af8a188bd08a9afb1b70c0cd055cea23

Change-Id: I012ff592004e9d372ad2264f56508ec893c9a9d0