Drop gunicorn from requirementes.txt
We dont ever actually import this anywhere or depend on it as a python
dependency.  Instead, its part of our diskimage-builder elements and
is largely an opinionated deployer choice we've made.  The only purpose
having it in requirements.txt is to pull it in during installation, which
should be delegated to ansible/DIB instead.

This also manually sync's global-requirements along with the update.

Closes-bug: #1526527

Change-Id: I834efb47ccda02a5163c5083836ff29fdc3fdd6f