Ensure interface cache up to date
The appliance server parses and caches the systems network interfaces
the first time it updates them, and never refreshes the cache. When
a new router interface is added, the appliance errors because its
interface cache has no idea about the NIC that corresponds to the
router interface.  This ensures we recreate this mapping anytime we
need it.

Depends-On: I4de32ba5e9e4f1ec479cb8763af55c2b2fa37da2

Closes-bug: #1531651
(cherry picked from commit adeaab79c83010d379779e2e33288cffc79e0ee1)

Change-Id: Iaff5a84a674d9089447bbdc8dc471f3d75a79af6