Adds keepalived based VRRPIPManager
This adds a new IP manager driver for configuring addresses
and routes via keepalived instead of directly.  It used when
the logical resource is configured to be highly-available,
according to configuration pushed by the orchestrator.

We rely on a 'ha_resource' flag attached to the main config
dict to enable it, and use specific HA config about peers and
cluster priority contained in the 'ha_config' section of the
main config.

The resulting keepalived cluster contains a VRRP instance for
each interface, with the exception of the management interface.

Change-Id: I5ababa41d65642b00f6b808197af9b2a59ebc67a
Depends-On: Ib679453aafe68e6653c8c5f9f412efa72c2b7cb1