Fix permissions on /etc/nginx/sites-enabled
With the switch to rootwrap, the API service now runs as the gunicorn
user but /etc/nginx/sites-enabled is still owned by root.  This updates
the DIB element to ensure its writable by gunicorn for config rendering.

Also makes a trivial update to releasenotes to remove the UNRELEASED
flag from mitaka.

Change-Id: Ieac128e47a44dd48acd00f68cd8e3a9ca15441ec
Closes-bug: #1558577